+++ enemies/Normie.java
@@ -7,6 +7,12 @@ public class Normie extends Enemy {
 
     public Normie(Vector2 pos) {
         super(pos, Color.red, 20);
+        this.health = 10;
+        this.damage = 1;
+
+        setAngle(45);
+        this.maxJumpStrength = 500;
+        this.jumpCooldown = 2;
     }
 
 }