+++ CHANGELOG.md
@@ -64,3 +64,41 @@ Methods
 - public void remove() // remove from Enemy list and handler objects
 
 [class Normie extends Enemy] added.
+
+
+### First Enemy AI
+
+Implemented new attributes for [abstract class Enemy]
+
+- public boolean airborne;
+- public double jumpCooldown;
+- public double jumpTimer;
+- public int maxJumpStrenght;
+- public int jumpAngleDegrees;
+- public double jumpAngleRadians;
+
+Implemented new method [public void setAngle(int angleDegrees)] as a constructor helper
+
+[void updateTimers(double dt)] now updates the enemy [jumpTimer], decreasing it to 0 while not being [airBorne]. Called in [void update(double dt)].
+
+Inside [update(double dt)], if [jumpTimer] reaches 0, the enemy will run [jumpTowardsTarget(pathToPlayer())] and set [jumpTimer] to [jumpCooldown].
+
+[public double pathToPlayer()] will return the angle towards the player, based on [jumpAngleRadians], and set [Vector2 target] to the players position. 
+
+[public void jumpTowardsTarget] will run the parabolic throw formula: 
+
+- if (difference != 0 && cos != 0) {
+            jumpVelocity = Math.sqrt(Math.abs((handler.gravity * dx * dx) / (2 * cos * cos * difference)));
+        }
+
+If the condition doesn't meet, the jumpVelocity will be set to [maxJumpStrength], which will also cap it. 
+
+[jumpTowardsTarget()] sets [airBorne] to true, which will be set to false once the [class LandingListener extends CollisionListener] gets triggered. 
+
+Added a few enemy types as examples:
+
+- Normie: red, 10 health,  20 radius, 45º jump, 500 max jump, 2s jump cd
+- Speedster: green, 10 health,  15 radius, 30º jump, 1000 max jump, 0.5s jump cd
+- Jumper: blue, 10 health,  25 radius, 70º jump, 1500 max jump, 2.5s jump cd
+
+[class Panel] now allows to accelerate [class PhysicsHandler] [anchorVelocity] inside [class MyKeyAdapter].